requirements-nn.txt
pytorch==1.12.1
tensorboard==2.10.1
